# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExtractInput {
  url: string;
}

export interface StreamLink {
  url: string;
  type: string;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  label?: string | null;
}

export interface ExtractResult {
  links: StreamLink[];
  pageTitle: string;
  sourceUrl: string;
  extractedAt: string;
  /** @nullable */
  error?: string | null;
}

export interface HistoryEntry {
  id: number;
  sourceUrl: string;
  pageTitle: string;
  linkCount: number;
  extractedAt: string;
}

/**
 * HTTP method cho API (GET hoặc POST)
 */
export type ChannelMethod = typeof ChannelMethod[keyof typeof ChannelMethod];


export const ChannelMethod = {
  GET: 'GET',
  POST: 'POST',
} as const;

/**
 * Cách đính kèm token vào request
 */
export type ChannelTokenType = typeof ChannelTokenType[keyof typeof ChannelTokenType];


export const ChannelTokenType = {
  bearer: 'bearer',
  cookie: 'cookie',
  query: 'query',
} as const;

export interface Channel {
  id: number;
  name: string;
  /** Trang web chứa player (dùng yt-dlp) */
  url: string;
  /** API endpoint trực tiếp để lấy link stream */
  apiUrl: string;
  /** HTTP method cho API (GET hoặc POST) */
  method: ChannelMethod;
  /** JSON string của custom HTTP headers */
  headers: string;
  notes: string;
  /** Endpoint để đăng nhập và lấy token */
  loginUrl: string;
  /** JSON body gửi khi login (dùng {username} và {password} làm placeholder) */
  loginBody: string;
  /** Tên đăng nhập */
  loginUsername: string;
  /** Mật khẩu */
  loginPassword: string;
  /** Đường dẫn tới token trong response JSON (VD data.token hoặc access_token) */
  tokenPath: string;
  /** Cách đính kèm token vào request */
  tokenType: ChannelTokenType;
  createdAt: string;
  updatedAt: string;
}

export type ChannelInputMethod = typeof ChannelInputMethod[keyof typeof ChannelInputMethod];


export const ChannelInputMethod = {
  GET: 'GET',
  POST: 'POST',
} as const;

export type ChannelInputTokenType = typeof ChannelInputTokenType[keyof typeof ChannelInputTokenType];


export const ChannelInputTokenType = {
  bearer: 'bearer',
  cookie: 'cookie',
  query: 'query',
} as const;

export interface ChannelInput {
  name: string;
  url?: string;
  apiUrl?: string;
  method?: ChannelInputMethod;
  headers?: string;
  notes?: string;
  loginUrl?: string;
  loginBody?: string;
  loginUsername?: string;
  loginPassword?: string;
  tokenPath?: string;
  tokenType?: ChannelInputTokenType;
}

export type ChannelInputPartialMethod = typeof ChannelInputPartialMethod[keyof typeof ChannelInputPartialMethod];


export const ChannelInputPartialMethod = {
  GET: 'GET',
  POST: 'POST',
} as const;

export type ChannelInputPartialTokenType = typeof ChannelInputPartialTokenType[keyof typeof ChannelInputPartialTokenType];


export const ChannelInputPartialTokenType = {
  bearer: 'bearer',
  cookie: 'cookie',
  query: 'query',
} as const;

export interface ChannelInputPartial {
  name?: string;
  url?: string;
  apiUrl?: string;
  method?: ChannelInputPartialMethod;
  headers?: string;
  notes?: string;
  loginUrl?: string;
  loginBody?: string;
  loginUsername?: string;
  loginPassword?: string;
  tokenPath?: string;
  tokenType?: ChannelInputPartialTokenType;
}

export interface DeleteResult {
  success: boolean;
}

export interface ApiError {
  error: string;
}