# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelMethod } from './channelMethod';
import type { ChannelTokenType } from './channelTokenType';

export interface Channel {
  id: number;
  name: string;
  /** Trang web chứa player (dùng yt-dlp) */
  url: string;
  /** API endpoint trực tiếp để lấy link stream */
  apiUrl: string;
  /** HTTP method cho API (GET hoặc POST) */
  method: ChannelMethod;
  /** JSON string của custom HTTP headers */
  headers: string;
  notes: string;
  /** Endpoint để đăng nhập và lấy token */
  loginUrl: string;
  /** JSON body gửi khi login (dùng {username} và {password} làm placeholder) */
  loginBody: string;
  /** Tên đăng nhập */
  loginUsername: string;
  /** Mật khẩu */
  loginPassword: string;
  /** Đường dẫn tới token trong response JSON (VD data.token hoặc access_token) */
  tokenPath: string;
  /** Cách đính kèm token vào request */
  tokenType: ChannelTokenType;
  createdAt: Date;
  updatedAt: Date;
}