# lib/api-zod/src/generated/types/channelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelInputMethod } from './channelInputMethod';
import type { ChannelInputTokenType } from './channelInputTokenType';

export interface ChannelInput {
  name: string;
  url?: string;
  apiUrl?: string;
  method?: ChannelInputMethod;
  headers?: string;
  notes?: string;
  loginUrl?: string;
  loginBody?: string;
  loginUsername?: string;
  loginPassword?: string;
  tokenPath?: string;
  tokenType?: ChannelInputTokenType;
}