# lib/api-zod/src/generated/types/channelInputPartial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelInputPartialMethod } from './channelInputPartialMethod';
import type { ChannelInputPartialTokenType } from './channelInputPartialTokenType';

export interface ChannelInputPartial {
  name?: string;
  url?: string;
  apiUrl?: string;
  method?: ChannelInputPartialMethod;
  headers?: string;
  notes?: string;
  loginUrl?: string;
  loginBody?: string;
  loginUsername?: string;
  loginPassword?: string;
  tokenPath?: string;
  tokenType?: ChannelInputPartialTokenType;
}