# lib/api-zod/src/generated/types/extractResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StreamLink } from './streamLink';

export interface ExtractResult {
  links: StreamLink[];
  pageTitle: string;
  sourceUrl: string;
  extractedAt: Date;
  /** @nullable */
  error?: string | null;
}