# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  sourceUrl: string;
  pageTitle: string;
  linkCount: number;
  extractedAt: Date;
}