# lib/api-zod/src/generated/types/streamLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StreamLink {
  url: string;
  type: string;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  label?: string | null;
}